updated return of service methods to optional

+++ src/main/java/com/sda/course/project/restaurant/validation/customExceptions/MealNotFoundException.java
@@ -1,7 +1,7 @@
 package com.sda.course.project.restaurant.validation.customExceptions;
 
 public class MealNotFoundException extends RuntimeException {
-    public MealNotFoundException(Long id){
+    public MealNotFoundException(Integer id){
         super("Meal id not found : " + id);
     }
 }
