+++ src/main/java/com/sda/course/project/restaurant/validation/customExceptions/OrderDetailNotFound.java
@@ -1,7 +1,7 @@
 package com.sda.course.project.restaurant.validation.customExceptions;
 
 public class OrderDetailNotFound extends RuntimeException {
-    public OrderDetailNotFound(Long id){
+    public OrderDetailNotFound(Integer id){
         super("Order detail id not found : " + id);
     }
 }
