+++ src/main/java/com/sda/course/project/restaurant/validation/customExceptions/PersonNotFoundException.java
@@ -1,7 +1,7 @@
 package com.sda.course.project.restaurant.validation.customExceptions;
 
 public class PersonNotFoundException extends RuntimeException {
-    public PersonNotFoundException(Long id){
+    public PersonNotFoundException(Integer id){
         super("Person id not found : " + id);
     }
 }
