+++ src/main/java/com/sda/course/project/restaurant/validation/customExceptions/RoleNotFoundException.java
@@ -1,7 +1,7 @@
 package com.sda.course.project.restaurant.validation.customExceptions;
 
 public class RoleNotFoundException extends RuntimeException {
-    public RoleNotFoundException (Long id){
+    public RoleNotFoundException (Integer id){
         super("Role id not found : " + id);
     }
 }
